requirements.txt
opencv-python
numpy
onvif-zeep
